requirements.txt
asgiref==3.4.1
cmdstanpy==0.9.68
convertdate==2.3.2
cycler==0.11.0
Cython==0.29.24
dj-database-url==0.5.0
Django==3.2.8
django-crispy-forms==1.13.0
django-heroku==0.0.0
django-pandas==0.6.5
ephem==4.1
gunicorn==20.1.0
hijri-converter==2.2.2
holidays==0.11.3.1
kiwisolver==1.3.2
korean-lunar-calendar==0.2.1
LunarCalendar==0.0.9
matplotlib==3.4.3
mysql-connector-python==8.0.27
mysqlclient==2.0.3
numpy==1.21.3
pandas==1.3.4
Pillow==8.4.0
prophet==1.0.1
protobuf==3.18.1
psycopg2-binary==2.9.2
PyMeeus==0.5.11
pyparsing==3.0.6
pystan==2.19.1.1
python-dateutil==2.8.2
python-decouple==3.5
pytz==2021.3
setuptools-git==1.2
six==1.16.0
sqlparse==0.4.2
tqdm==4.62.3
typing-extensions==3.10.0.2
ujson==4.2.0
whitenoise==5.3.0
